+++ Going.UI.Tests/GoStepBarTests.cs
@@ -1,5 +1,6 @@
 using System.Text.Json;
 using Going.UI.Controls;
+using Going.UI.Datas;
 using Going.UI.Json;
 using Xunit;